fix the kincso part
